chore: refactor testing
